LPS-111198 Inline CryptoHashFlavor
since we are no longer serializing it we don't need to have a separate
class for it.
