--- modules/apps/portal-crypto-hash/portal-crypto-hash-api/src/main/java/com/liferay/portal/crypto/hash/flavor/CryptoHashFlavor.java
@@ -1,41 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.crypto.hash.flavor;
-
-import java.util.Optional;
-
-/**
- * @author Arthur Chan
- * @author Carlos Sierra Andrés
- */
-public final class CryptoHashFlavor {
-
-	public CryptoHashFlavor(String pepperId, byte[] salt) {
-		_pepperId = pepperId;
-		_salt = salt;
-	}
-
-	public Optional<String> getPepperId() {
-		return Optional.ofNullable(_pepperId);
-	}
-
-	public Optional<byte[]> getSalt() {
-		return Optional.ofNullable(_salt);
-	}
-
-	private final String _pepperId;
-	private final byte[] _salt;
-
-}
