--- modules/apps/portal-crypto-hash/portal-crypto-hash-api/src/main/resources/com/liferay/portal/crypto/hash/flavor/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
